feat(api): More filters + refactoring
