+++ api/src/lib.rs
@@ -1,1 +1,3 @@
+pub mod context;
+pub mod schema;
 pub mod query_mapping;
