+++ api/src/main.rs
@@ -7,601 +7,16 @@ use axum::{
     routing::{get, on, MethodFilter},
     Extension, Router,
 };
-use chrono::{DateTime, Utc};
 use clap::{Args, Parser};
 use juniper::{
-    graphql_object, EmptyMutation, EmptySubscription, Executor, GraphQLEnum, GraphQLInputObject,
-    GraphQLObject, RootNode, ScalarValue,
+    EmptyMutation, EmptySubscription, RootNode
 };
 use juniper_axum::{extract::JuniperRequest, graphiql, playground, response::JuniperResponse};
-use sdk::{mapping, system_ids};
+use sdk::neo4rs;
 use tokio::net::TcpListener;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
-#[derive(Clone)]
-pub struct KnowledgeGraph(Arc<sink::kg::Client>);
-
-impl juniper::Context for KnowledgeGraph {}
-
-#[derive(Clone)]
-pub struct Query;
-
-#[graphql_object]
-#[graphql(context = KnowledgeGraph, scalar = S: ScalarValue)]
-impl Query {
-    /// Returns a single entity identified by its ID and space ID
-    async fn entity<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-        id: String,
-        space_id: String,
-        // version_id: Option<String>,
-    ) -> Option<Entity> {
-        // let query = QueryMapper::default().select_root_node(&id, &executor.look_ahead()).build();
-        // tracing::info!("Query: {}", query);
-
-        mapping::Entity::<mapping::Triples>::find_by_id(&executor.context().0.neo4j, &id, &space_id)
-            .await
-            .expect("Failed to find entity")
-            .map(Entity::from)
-    }
-
-    /// Returns multiple entities according to the provided space ID and filter
-    async fn entities<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-        r#where: Option<EntityWhereFilter>,
-    ) -> Vec<Entity> {
-        // let query = QueryMapper::default().select_root_node(&id, &executor.look_ahead()).build();
-        // tracing::info!("Query: {}", query);
-
-        match r#where {
-            Some(r#where) => mapping::Entity::<mapping::Triples>::find_many(
-                &executor.context().0.neo4j,
-                Some(r#where.into()),
-            )
-            .await
-            .expect("Failed to find entities")
-            .into_iter()
-            .map(Entity::from)
-            .collect::<Vec<_>>(),
-            _ => mapping::Entity::<mapping::Triples>::find_many(&executor.context().0.neo4j, None)
-                .await
-                .expect("Failed to find entities")
-                .into_iter()
-                .map(Entity::from)
-                .collect::<Vec<_>>(),
-        }
-    }
-
-    /// Returns a single relation identified by its ID and space ID
-    async fn relation<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-        id: String,
-        space_id: String,
-        // version_id: Option<String>,
-    ) -> Option<Relation> {
-        mapping::Relation::<mapping::Triples>::find_by_id(
-            &executor.context().0.neo4j,
-            &id,
-            &space_id,
-        )
-        .await
-        .expect("Failed to find relation")
-        .map(|rel| rel.into())
-    }
-
-    /// Returns multiple relations according to the provided space ID and filter
-    async fn relations<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-        space_id: String,
-        // version_id: Option<String>,
-        filter: Option<RelationFilter>,
-    ) -> Vec<Relation> {
-        match filter {
-            Some(RelationFilter {
-                relation_types: Some(types),
-            }) if !types.is_empty() => mapping::Relation::<mapping::Triples>::find_by_types(
-                &executor.context().0.neo4j,
-                &types,
-                &space_id,
-            )
-            .await
-            .expect("Failed to find relations")
-            .into_iter()
-            .map(|rel| rel.into())
-            .collect::<Vec<_>>(),
-            _ => mapping::Relation::<mapping::Triples>::find_all(
-                &executor.context().0.neo4j,
-                &space_id,
-            )
-            .await
-            .expect("Failed to find relations")
-            .into_iter()
-            .map(|rel| rel.into())
-            .collect::<Vec<_>>(),
-        }
-    }
-}
-
-/// Entity filter input object
-///
-/// ```graphql
-/// query {
-///     entities(where: {
-///         space_id: "BJqiLPcSgfF8FRxkFr76Uy",
-///         types_contain: ["XG26vy98XAA6cR6DosTALk", "XG26vy98XAA6cR6DosTALk"],
-///         attributes_contain: [
-///             {id: "XG26vy98XAA6cR6DosTALk", value: "value", value_type: TEXT},
-///         ]
-///     })
-/// }
-/// ```
-///
-#[derive(Debug, GraphQLInputObject)]
-struct EntityFilter {
-    /// Filter by entity types
-    r#where: Option<EntityWhereFilter>,
-}
-
-#[derive(Debug, GraphQLInputObject)]
-struct EntityWhereFilter {
-    id: Option<String>,
-    space_id: Option<String>,
-    types_contain: Option<Vec<String>>,
-    attributes_contain: Option<Vec<EntityAttributeFilter>>,
-}
-
-impl From<EntityWhereFilter> for mapping::EntityWhereFilter {
-    fn from(filter: EntityWhereFilter) -> Self {
-        mapping::EntityWhereFilter {
-            id: filter.id,
-            space_id: filter.space_id,
-            types_contain: filter.types_contain,
-            attributes_contain: filter
-                .attributes_contain
-                .map(|filters| filters.into_iter().map(Into::into).collect()),
-        }
-    }
-}
-
-#[derive(Debug, GraphQLInputObject)]
-struct EntityAttributeFilter {
-    attribute: String,
-    value: Option<String>,
-    value_type: Option<ValueType>,
-}
-
-impl From<EntityAttributeFilter> for mapping::EntityAttributeFilter {
-    fn from(filter: EntityAttributeFilter) -> Self {
-        mapping::EntityAttributeFilter {
-            attribute: filter.attribute,
-            value: filter.value,
-            value_type: filter.value_type.map(Into::into),
-        }
-    }
-}
-
-/// Relation filter input object
-#[derive(Debug, GraphQLInputObject)]
-struct RelationFilter {
-    /// Filter by relation types
-    relation_types: Option<Vec<String>>,
-}
-
-#[derive(Debug)]
-pub struct Entity {
-    id: String,
-    types: Vec<String>,
-    space_id: String,
-    created_at: DateTime<Utc>,
-    created_at_block: String,
-    updated_at: DateTime<Utc>,
-    updated_at_block: String,
-    attributes: Vec<Triple>,
-}
-
-impl From<mapping::Entity<mapping::Triples>> for Entity {
-    fn from(entity: mapping::Entity<mapping::Triples>) -> Self {
-        Self {
-            id: entity.attributes.id,
-            types: entity.types,
-            space_id: entity.attributes.system_properties.space_id.clone(),
-            created_at: entity.attributes.system_properties.created_at,
-            created_at_block: entity.attributes.system_properties.created_at_block,
-            updated_at: entity.attributes.system_properties.updated_at,
-            updated_at_block: entity.attributes.system_properties.updated_at_block,
-            attributes: entity
-                .attributes
-                .attributes
-                .into_iter()
-                .map(|(key, triple)| Triple {
-                    space_id: entity.attributes.system_properties.space_id.clone(),
-                    attribute: key,
-                    value: triple.value,
-                    value_type: triple.value_type.into(),
-                    options: Options {
-                        format: triple.options.format,
-                        unit: triple.options.unit,
-                        language: triple.options.language,
-                    },
-                })
-                .collect(),
-        }
-    }
-}
-
-#[graphql_object]
-#[graphql(context = KnowledgeGraph, scalar = S: ScalarValue)]
-/// Entity object
-impl Entity {
-    /// Entity ID
-    fn id(&self) -> &str {
-        &self.id
-    }
-
-    /// Entity name (if available)
-    fn name(&self) -> Option<&str> {
-        self.attributes
-            .iter()
-            .find(|triple| triple.attribute == system_ids::NAME)
-            .map(|triple| triple.value.as_str())
-    }
-
-    /// The space ID of the entity (note: the same entity can exist in multiple spaces)
-    fn space_id(&self) -> &str {
-        &self.space_id
-    }
-
-    fn created_at(&self) -> String {
-        self.created_at.to_rfc3339()
-    }
-
-    fn created_at_block(&self) -> &str {
-        &self.created_at_block
-    }
-
-    fn updated_at(&self) -> String {
-        self.updated_at.to_rfc3339()
-    }
-
-    fn updated_at_block(&self) -> &str {
-        &self.updated_at_block
-    }
-
-    /// Types of the entity (which are entities themselves)
-    async fn types<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-    ) -> Vec<Entity> {
-        if self.types.contains(&system_ids::RELATION_TYPE.to_string()) {
-            // Since relations are also entities, and a relation's types are modelled differently
-            // in Neo4j, we need to check fetch types differently if the entity is a relation.
-            // mapping::Relation::<mapping::Triples>::find_types(
-            //     &executor.context().0.neo4j,
-            //     &self.id,
-            //     &self.space_id,
-            // )
-            // .await
-            // .expect("Failed to find relations")
-            // .into_iter()
-            // .map(|rel| rel.into())
-            // .collect::<Vec<_>>()
-
-            // For now, we'll just return the relation type
-            mapping::Entity::<mapping::Triples>::find_by_id(
-                &executor.context().0.neo4j,
-                system_ids::RELATION_TYPE,
-                &self.space_id,
-            )
-            .await
-            .expect("Failed to find types")
-            .map(|rel| vec![rel.into()])
-            .unwrap_or(vec![])
-        } else {
-            mapping::Entity::<mapping::Triples>::find_types(
-                &executor.context().0.neo4j,
-                &self.id,
-                &self.space_id,
-            )
-            .await
-            .expect("Failed to find relations")
-            .into_iter()
-            .map(|rel| rel.into())
-            .collect::<Vec<_>>()
-        }
-    }
-
-    /// Attributes of the entity
-    fn attributes(&self, filter: Option<AttributeFilter>) -> Vec<&Triple> {
-        match filter {
-            Some(AttributeFilter {
-                value_type: Some(value_type),
-            }) => self
-                .attributes
-                .iter()
-                .filter(|triple| triple.value_type == value_type)
-                .collect::<Vec<_>>(),
-            _ => self.attributes.iter().collect::<Vec<_>>(),
-        }
-    }
-
-    /// Relations outgoing from the entity
-    async fn relations<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-    ) -> Vec<Relation> {
-        mapping::Entity::<mapping::Triples>::find_relations::<mapping::Triples>(
-            &executor.context().0.neo4j,
-            &self.id,
-            &self.space_id,
-        )
-        .await
-        .expect("Failed to find relations")
-        .into_iter()
-        .map(|rel| rel.into())
-        .collect::<Vec<_>>()
-    }
-}
-
-impl From<mapping::ValueType> for ValueType {
-    fn from(value_type: mapping::ValueType) -> Self {
-        match value_type {
-            mapping::ValueType::Text => Self::Text,
-            mapping::ValueType::Number => Self::Number,
-            mapping::ValueType::Checkbox => Self::Checkbox,
-            mapping::ValueType::Url => Self::Url,
-            mapping::ValueType::Time => Self::Time,
-            mapping::ValueType::Point => Self::Point,
-        }
-    }
-}
-
-impl From<ValueType> for mapping::ValueType {
-    fn from(value_type: ValueType) -> Self {
-        match value_type {
-            ValueType::Text => mapping::ValueType::Text,
-            ValueType::Number => mapping::ValueType::Number,
-            ValueType::Checkbox => mapping::ValueType::Checkbox,
-            ValueType::Url => mapping::ValueType::Url,
-            ValueType::Time => mapping::ValueType::Time,
-            ValueType::Point => mapping::ValueType::Point,
-        }
-    }
-}
-
-#[derive(Debug, GraphQLInputObject)]
-struct AttributeFilter {
-    value_type: Option<ValueType>,
-}
-
-#[derive(Debug)]
-pub struct Relation {
-    id: String,
-    relation_types: Vec<String>,
-    space_id: String,
-    created_at: DateTime<Utc>,
-    created_at_block: String,
-    updated_at: DateTime<Utc>,
-    updated_at_block: String,
-    attributes: Vec<Triple>,
-}
-
-impl From<mapping::Relation<mapping::Triples>> for Relation {
-    fn from(relation: mapping::Relation<mapping::Triples>) -> Self {
-        Self {
-            id: relation.attributes.id,
-            relation_types: relation.types,
-            space_id: relation.attributes.system_properties.space_id.clone(),
-            created_at: relation.attributes.system_properties.created_at,
-            created_at_block: relation
-                .attributes
-                .system_properties
-                .created_at_block
-                .clone(),
-            updated_at: relation.attributes.system_properties.updated_at,
-            updated_at_block: relation
-                .attributes
-                .system_properties
-                .updated_at_block
-                .clone(),
-            attributes: relation
-                .attributes
-                .attributes
-                .iter()
-                .map(|(key, triple)| Triple {
-                    // entiti: triple.entity,
-                    space_id: relation.attributes.system_properties.space_id.clone(),
-                    attribute: key.to_string(),
-                    value: triple.value.clone(),
-                    value_type: triple.value_type.clone().into(),
-                    options: Options {
-                        format: triple.options.format.clone(),
-                        unit: triple.options.unit.clone(),
-                        language: triple.options.language.clone(),
-                    },
-                })
-                .collect(),
-        }
-    }
-}
-
-#[graphql_object]
-#[graphql(context = KnowledgeGraph, scalar = S: ScalarValue)]
-/// Relation object
-///
-/// Note: Relations are also entities, but they have a different structure in the database.
-/// In other words, the Relation object is a "view" on a relation entity. All relations
-/// can also be queried as entities.
-impl Relation {
-    /// Relation ID
-    fn id(&self) -> &str {
-        &self.id
-    }
-
-    /// Relation name (if available)
-    fn name(&self) -> Option<&str> {
-        self.attributes
-            .iter()
-            .find(|triple| triple.attribute == system_ids::NAME)
-            .map(|triple| triple.value.as_str())
-    }
-
-    fn created_at(&self) -> String {
-        self.created_at.to_rfc3339()
-    }
-
-    fn created_at_block(&self) -> &str {
-        &self.created_at_block
-    }
-
-    fn updated_at(&self) -> String {
-        self.updated_at.to_rfc3339()
-    }
-
-    fn updated_at_block(&self) -> &str {
-        &self.updated_at_block
-    }
-
-    /// Attributes of the relation
-    fn attributes(&self) -> &[Triple] {
-        &self.attributes
-    }
-
-    /// Relation types of the relation
-    async fn relation_types<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-    ) -> Vec<Entity> {
-        mapping::Entity::<mapping::Triples>::find_by_ids(
-            &executor.context().0.neo4j,
-            &self.relation_types,
-            &self.space_id,
-        )
-        .await
-        .expect("Failed to find types")
-        .into_iter()
-        .filter(|rel| rel.id() != system_ids::RELATION_TYPE)
-        .map(|rel| rel.into())
-        .collect::<Vec<_>>()
-    }
-
-    /// Entity from which the relation originates
-    async fn from<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-    ) -> Entity {
-        mapping::Relation::<mapping::Triples>::find_from::<mapping::Triples>(
-            &executor.context().0.neo4j,
-            &self.id,
-            &self.space_id,
-        )
-        .await
-        .expect("Failed to find node")
-        .map(Entity::from)
-        .unwrap()
-    }
-
-    /// Entity to which the relation points
-    async fn to<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-    ) -> Entity {
-        mapping::Relation::<mapping::Triples>::find_to::<mapping::Triples>(
-            &executor.context().0.neo4j,
-            &self.id,
-            &self.space_id,
-        )
-        .await
-        .expect("Failed to find node")
-        .map(Entity::from)
-        .unwrap()
-    }
-
-    /// Relations outgoing from the relation
-    async fn relations<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-    ) -> Vec<Relation> {
-        mapping::Entity::<mapping::Triples>::find_relations::<mapping::Triples>(
-            &executor.context().0.neo4j,
-            &self.id,
-            &self.space_id,
-        )
-        .await
-        .expect("Failed to find relations")
-        .into_iter()
-        .map(|rel| rel.into())
-        .collect::<Vec<_>>()
-    }
-}
-
-#[derive(Debug)]
-struct Triple {
-    space_id: String,
-    attribute: String,
-    value: String,
-    value_type: ValueType,
-    options: Options,
-}
-
-#[graphql_object]
-#[graphql(context = KnowledgeGraph, scalar = S: ScalarValue)]
-impl Triple {
-    /// Attribute ID of the triple
-    fn attribute(&self) -> &str {
-        &self.attribute
-    }
-
-    /// Value of the triple
-    fn value(&self) -> &str {
-        &self.value
-    }
-
-    /// Value type of the triple
-    fn value_type(&self) -> &ValueType {
-        &self.value_type
-    }
-
-    /// Options of the triple (if any)
-    fn options(&self) -> &Options {
-        &self.options
-    }
-
-    /// Name of the attribute (if available)
-    async fn name<'a, S: ScalarValue>(
-        &'a self,
-        executor: &'a Executor<'_, '_, KnowledgeGraph, S>,
-    ) -> Option<String> {
-        mapping::Entity::<mapping::Named>::find_by_id(
-            &executor.context().0.neo4j,
-            &self.attribute,
-            &self.space_id,
-        )
-        .await
-        .expect("Failed to find attribute entity")
-        .and_then(|entity| entity.name())
-    }
-}
-
-#[derive(Debug, GraphQLEnum, PartialEq)]
-pub enum ValueType {
-    Text,
-    Number,
-    Checkbox,
-    Url,
-    Time,
-    Point,
-}
-
-#[derive(Debug, GraphQLObject)]
-struct Options {
-    pub format: Option<String>,
-    pub unit: Option<String>,
-    pub language: Option<String>,
-}
+use api::{context::KnowledgeGraph, schema::Query};
 
 type Schema =
     RootNode<'static, Query, EmptyMutation<KnowledgeGraph>, EmptySubscription<KnowledgeGraph>>;
@@ -621,7 +36,7 @@ async fn main() -> anyhow::Result<()> {
 
     let args = AppArgs::parse();
 
-    let kg_client = sink::kg::Client::new(
+    let neo4j = neo4rs::Graph::new(
         &args.neo4j_args.neo4j_uri,
         &args.neo4j_args.neo4j_user,
         &args.neo4j_args.neo4j_pass,
@@ -647,7 +62,7 @@ async fn main() -> anyhow::Result<()> {
         .route("/playground", get(playground("/graphql", "/subscriptions")))
         .route("/", get(homepage))
         .layer(Extension(Arc::new(schema)))
-        .layer(Extension(KnowledgeGraph(Arc::new(kg_client))));
+        .layer(Extension(KnowledgeGraph::new(Arc::new(neo4j))));
 
     let addr = SocketAddr::from(([0, 0, 0, 0], 8080));
     let listener = TcpListener::bind(addr)
