+++ api/src/schema/options.rs
@@ -1,0 +1,9 @@
+use juniper::GraphQLObject;
+
+
+#[derive(Debug, GraphQLObject)]
+pub struct Options {
+    pub(crate) format: Option<String>,
+    pub(crate) unit: Option<String>,
+    pub(crate) language: Option<String>,
+}
