+++ api/src/schema/relation_filter.rs
@@ -1,0 +1,8 @@
+use juniper::GraphQLInputObject;
+
+/// Relation filter input object
+#[derive(Debug, GraphQLInputObject)]
+pub struct RelationFilter {
+    /// Filter by relation types
+    pub(crate) relation_types: Option<Vec<String>>,
+}
