+++ api/src/schema/value_type.rs
@@ -1,0 +1,11 @@
+use juniper::GraphQLEnum;
+
+#[derive(Debug, GraphQLEnum, PartialEq)]
+pub enum ValueType {
+    Text,
+    Number,
+    Checkbox,
+    Url,
+    Time,
+    Point,
+}
