+++ sdk/src/mapping/mod.rs
@@ -3,11 +3,13 @@ pub mod entity;
 pub mod entity_filter;
 pub mod query;
 pub mod relation;
+pub mod relation_filter;
 pub mod triple;
 
 pub use attributes::Attributes;
 pub use entity::{Entity, Named};
-pub use entity_filter::{EntityAttributeFilter, EntityWhereFilter, EntityWhereFilterBuilder};
+pub use entity_filter::{EntityAttributeFilter, EntityFilter, EntityRelationFilter};
 pub use query::Query;
 pub use relation::Relation;
+pub use relation_filter::RelationFilter;
 pub use triple::{Options, Triple, Triples, ValueType};
